+++ src/site/twitch.tv/modules/chat-input/ChatInput.vue
@@ -1,3 +1,5 @@
+<template />
+
 <!-- eslint-disable prettier/prettier -->
 <script setup lang="ts">
 import { onMounted, onUnmounted, ref, watch } from "vue";