+++ src/types/vite-env.d.ts
@@ -7,3 +7,8 @@ declare module "*.vue" {
 	const component: DefineComponent<Record<string, never>, Record<string, never>, any>;
 	export default component;
 }
+
+declare module "*?sharedworker&inline" {
+	const WorkerFactory: new () => SharedWorker;
+	export default WorkerFactory;
+}