+++ src/workers.d.ts
@@ -1,4 +1,0 @@
-declare module "*?sharedworker&inline" {
-	const WorkerFactory: new () => SharedWorker;
-	export default WorkerFactory;
-}