+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### Version 3.0.0.4000 (Beta 4)
+
+- The emote menu button will now take the form of the 7TV Logo
+- Fixed an issue which caused this changelog to become impossible to close if multiple tabs were open
+
 ### Version 3.0.0.3000 (Beta 3)
 
 - Added an option for alternating background colors between chat messages